refactor canvas selection logic
